backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
asyncpg==0.28.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pandas==2.0.3
numpy==1.25.2